BISERVER-3170: Fixed problem with ComponentBase.
What was happening here is that when ComponentBase.getComponentSetting(final String path) was being called.  It would check it's instance variable "componentDefinitionMap" to see if the path was defined there.  The problem is that ComponentBase doesn't actually instantiate a componentDefinitionMap directly.  It relies on the componentDefinitionMap setter (in this case SQLBaseComponent).  However, if the subclass never creates one then the componentDefinitionMap will throw a null pointer when calling it's get(path).  Added a check to determine if this object is null before trying to execute the get(path).  I could have fixed SQLBaseComponent to create a componentDefitionMap but it's possible this is occurring in other plugins.

git-svn-id: http://source.pentaho.org/svnroot/bi-platform-v2/trunk@38390 5d404267-db13-0410-8476-b1f6b398e63a