BISERVER-2588: At first I couldn't get this to reproduce.  This is actually a semi-user error.
When there is a large amount of content items the deletion of a subscription (public schedule) can take several seconds.  Since the GUI doesn't provide any feedback after the delete is hit the user doesn't know the delete has been started and they hit delete a second time.  Since the first transaction is already in the middle of it's delete hibernate throws the stack trace seen above along with the spurious error in the PUC.  Even though everything worked as it was suppose to the user still gets the error.

The fix is to remove the delete label after it has been pressed so that a second press isn't possible.

git-svn-id: http://source.pentaho.org/svnroot/bi-platform-v2/trunk@38797 5d404267-db13-0410-8476-b1f6b398e63a